Hack to make Sonarr prioritize preffered words
This change moves preffered words check up the decision tree when downloading
and upgrading. This is to prevent Sonaar from replacing dubbed shows
with subtitled ones when the subs releases were higher quality video.
This was specifically an issue with anime, where I value the right audio
vs better quality video.

For titles that don't involve the preferred words there isn't a
difference.

Signed-off-by: Josh Farwell <josh.farwell@gmail.com>